web/fastapi/fastapi_template/requirements.txt
# FastAPI и зависимости
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Конфигурация и валидация
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# База данных (опционально)
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1

# Логирование
loguru==0.7.2

# Аутентификация (опционально)
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Утилиты
python-multipart==0.0.6
email-validator==2.1.0

# Для разработки
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1
black==23.11.0
isort==5.12.0
mypy==1.7.0